Decouple breadcrumbs from hero component
