+++ src/components/breadcrumbs/index.ts
@@ -1,0 +1,3 @@
+import Breadcrumbs from './breadcrumbs';
+
+export default Breadcrumbs;
