+++ src/components/breadcrumbs/types.ts
@@ -1,0 +1,8 @@
+export interface Crumb {
+    text: string;
+    url?: string;
+}
+
+export interface BreadcrumbsProps {
+    crumbs: Crumb[];
+}
